[ feat ] 임시 알림 구현 위한 버튼 추가
